Update .gitignore and Dependencies.toml, enhance types.bal and mock_service.bal
- Updated .gitignore to exclude all jar files except for gradle-wrapper.jar.
- Downgraded the version of the 'observe' package in Dependencies.toml from 1.6.0 to 1.5.1.
- Added a comment in types.bal to clarify the purpose of the DIRECTORY enum value.
- Improved error handling in mock_service.bal by refining the logic for target file paths in hard link creation.

+++ native/src/main/java/io/ballerina/lib/azure/storage/files/AdminOps.java
@@ -77,7 +77,7 @@ public final class AdminOps {
                 sdkOptions.setMetadata(ValueUtils.optStringMap(record, Constants.METADATA));
                 Object quota = record.get(Constants.QUOTA_IN_GB);
                 if (quota != null) {
-                    sdkOptions.setQuotaInGb(((Long) quota).intValue());
+                    sdkOptions.setQuotaInGb(Math.toIntExact((Long) quota));
                 }
                 String tier = ValueUtils.optString(record, Constants.ACCESS_TIER);
                 if (tier != null) {
